add allocation bar to new cluster page
