+++ webui/react/src/pages/Cluster/ClusterHistoricalUsage.module.scss
@@ -3,8 +3,8 @@
   border-bottom: 1px solid var(--theme-colors-monochrome-12);
   left: 0;
   margin-bottom: var(--theme-sizes-layout-big);
-  margin-top: calc(var(--theme-sizes-layout-medium) * -2);
   padding-bottom: var(--theme-sizes-layout-medium);
+  padding-right: var(--theme-sizes-layout-big);
   padding-top: var(--theme-sizes-layout-medium);
   position: absolute;
   right: 0;
