+++ webui/react/src/pages/Cluster/ClusterOverview.tsx
@@ -3,7 +3,6 @@ import React, { useCallback, useEffect, useMemo, useState } from 'react';
 
 import Grid, { GridMode } from 'components/Grid';
 import GridListRadioGroup, { GridListView } from 'components/GridListRadioGroup';
-import Message, { MessageType } from 'components/Message';
 import OverviewStats from 'components/OverviewStats';
 import ResourcePoolCard from 'components/ResourcePoolCard';
 import ResourcePoolDetails from 'components/ResourcePoolDetails';
@@ -24,6 +23,7 @@ import {
 } from 'types';
 import { getSlotContainerStates } from 'utils/cluster';
 
+import { ClusterOverallBar } from './ClusterOverallBar';
 import css from './ClusterOverview.module.scss';
 
 const STORAGE_PATH = 'cluster';
@@ -69,18 +69,6 @@ const ClusterOverview: React.FC = () => {
     });
     return tally;
   }, [ resourcePools ]);
-
-  const cudaSlotStates = useMemo(() => {
-    return getSlotContainerStates(agents || [], ResourceType.CUDA);
-  }, [ agents ]);
-
-  const rocmSlotStates = useMemo(() => {
-    return getSlotContainerStates(agents || [], ResourceType.ROCM);
-  }, [ agents ]);
-
-  const cpuSlotStates = useMemo(() => {
-    return getSlotContainerStates(agents || [], ResourceType.CPU);
-  }, [ agents ]);
 
   const [ cudaTotalSlots, rocmTotalSlots ] = useMemo(() => {
     return resourcePools.reduce((acc, pool) => {
@@ -213,38 +201,7 @@ const ClusterOverview: React.FC = () => {
           ) : null}
         </Grid>
       </Section>
-      <Section hideTitle title="Overall Allocation">
-        {cudaTotalSlots + rocmTotalSlots + overview.CPU.total === 0 ? (
-          <Message title="No connected agents." type={MessageType.Empty} />
-        ) : null }
-        {cudaTotalSlots > 0 && (
-          <SlotAllocationBar
-            resourceStates={cudaSlotStates}
-            showLegends
-            size={ShirtSize.enormous}
-            title={`Compute (${ResourceType.CUDA})`}
-            totalSlots={cudaTotalSlots}
-          />
-        )}
-        {rocmTotalSlots > 0 && (
-          <SlotAllocationBar
-            resourceStates={rocmSlotStates}
-            showLegends
-            size={ShirtSize.enormous}
-            title={`Compute (${ResourceType.ROCM})`}
-            totalSlots={rocmTotalSlots}
-          />
-        )}
-        {overview.CPU.total > 0 && (
-          <SlotAllocationBar
-            resourceStates={cpuSlotStates}
-            showLegends
-            size={ShirtSize.enormous}
-            title={`Compute (${ResourceType.CPU})`}
-            totalSlots={overview.CPU.total}
-          />
-        )}
-      </Section>
+      <ClusterOverallBar />
       <Section
         options={<GridListRadioGroup value={selectedView} onChange={handleRadioChange} />}
         title={`${resourcePools.length} Resource Pools`}>
