+++ webui/react/src/pages/Clusters.module.scss
@@ -1,0 +1,3 @@
+.tab {
+  padding: 0 var(--theme-sizes-layout-big);
+}
